requirements.txt
appbuilder_sdk==0.8.0
cachetools==5.3.0
erniebot==0.5.5
erniebot_agent==0.5.2
gradio==4.26.0
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.7
librosa==0.10.2.post1
loguru==0.7.0
numpy==1.24.2
Requests==2.32.3
soundfile==0.12.1
